refactor: modularizacion y reestructuracion de directorios y funcionalidades del bot
