+++ Pipfile
@@ -24,4 +24,4 @@ Werkzeug = "==1.0.1"
 [dev-packages]
 
 [requires]
-python_version = "3.8"
+python_version = "3"
