+++ assets/messages/calendario-academico.txt
@@ -1,0 +1,21 @@
+<b><u>Cursada 2° Cuatrimestre 2021</u></b>
+<b>Regulares</b>
+Inscripcion: 02 al 04 de Agosto de 2021
+Inicio: 09 de Agosto de 2021
+Finalizacion: 27 de Noviembre de 2021
+
+<b>Ingresantes</b>
+Inscripcion: 19 y 20 de Julio de 2021
+Inicio: 26 de Julio de 2021
+Finalizacion: 27 de Noviembre de 2021
+
+<b><u>Inscripción a Carreras de Grado: 1º cuatrimestre de 2022</u></b>
+Inscripcion: 04 de Octubre al 27 de Noviembre de 2021
+
+<b><u>Exámenes Turno Diciembre 2021</u></b>
+Inscripcion: 30 de Noviembre al 01 de Diciembre de 2021
+Examanes: 06 al 18 de Diciembre de 2021
+
+<b><u>Exámenes Turno Febrero 2022</u></b>
+Inscripcion: 01 al 05 de Febrero de 2022
+Examanes: 14 al 26 de Febrero de 2022
