+++ assets/messages/help.txt
@@ -1,0 +1,11 @@
+<b>infoUNObot</b> te brinda información necesaria sobre la Universidad Nacional del Oeste, desde el estado del siu o el campus, hasta fechas importantes o links útiles.
+
+<b>/help</b> ó <b>/ayuda</b> - Muestra este mensaje.
+<b>/siu</b> - Obtiene información del siu para saber el estado del mismo y su latencia.
+<b>/campus</b> - Obtiene información del campus virtual para saber el estado del mismo y su latencia.
+<b>/links</b> - Te devuelve un listado de links útiles sobre la carrera (grupos, comunidades, etc).
+<b>/comunidades</b> ó <b>/comunidades_it</b> - Te devuelve un listado de links de nuestras comunidades.
+<b>/calendar</b> ó <b>/calendario</b> - Te muestra las fechas importantes del calendario académico de la Universidad. Si escribis <b>/calendar feriados</b> te muestra los feriados del año corriente.
+<b>/mails</b> - Te muestra los mails más importantes de las escuelas, además si especificás la escuela te filtra el resultado.
+
+Este bot fue posible y llevado a cabo gracias a GNUno, cualquier consulta o pregunta hacela aquí: <i>https://t.me/gnuno_merlo</i>.
