+++ bot.py
@@ -1,141 +1,22 @@
 import telebot
 from flask import Flask, request
 import os
-import logging
-import requests
-import time
-import messages as responses
-
-# Configurar logging
-logging.basicConfig(
-    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s', level=logging.INFO)
-logger = telebot.logger
-telebot.logger.setLevel(logging.DEBUG)
-
-MODE = os.environ.get("MODE")
+from src.setup import MODE, API_TOKEN, bot, logger
+from src.accepted_commands import accepted_commands, handlers
 
 if MODE == "prod":
-    API_TOKEN = os.environ.get('TOKEN')
-    bot = telebot.TeleBot(API_TOKEN, parse_mode='HTML')
-
     def run():
         server.run(host="0.0.0.0", port=int(os.environ.get('PORT', 5000)))
 else:
-
-    API_TOKEN = '' #Carga aqui el TOKEN de prueba
-
-    bot = telebot.TeleBot(API_TOKEN, parse_mode='HTML')
-
     def run():
         bot.polling()
 
 server = Flask(__name__)
 
 
-def get_info(url):
-    url_data = dict()
-    try:
-        t0 = time.time()
-        r = requests.get(url)
-        r.raise_for_status()
-        t1 = time.time()
-        url_data['status'] = r.status_code
-        url_data['latency'] = int((t1-t0) * 1000)
-        return url_data
-    except requests.exceptions.RequestException as err:
-        return url_data
-
-
-def get_links_from_api():
-    res = requests.get('https://igna98.alwaysdata.net/page')
-    data = res.json()['data']
-    mapped_list = list(map(lambda link: link['name'], data))
-    return responses.links_message(mapped_list)
-
-
-def url_message(url, name):
-    url_data = get_info(url)
-    if 'status' in url_data.keys() and url_data['status'] == 200:
-        return responses.url_success_message(url_data['latency'], name)
-    else:
-        return responses.url_failure_message(name)
-
-
-def get_links_message():
-    return get_links_from_api()
-
-
-@bot.message_handler(commands=['start'])
-def send_welcome(message):
-    bot.reply_to(
-        message, "Bienvenidx a este botardo con información util sobre la Universidad Nacional del Oeste. Escribí <b>/help</b> para saber cómo seguir.")
-
-
-@bot.message_handler(commands=['help','ayuda'])
-def help_message(message):
-    user_id = message.from_user.id
-    logger.info(f"El usuario {user_id} ha solicitado AYUDA.")
-    chat_id = message.chat.id
-    bot.send_message(chat_id, responses.help_message())
-
-
-@bot.message_handler(commands=['links'])
-def get_useful_links(message):
-    user_id = message.from_user.id
-    logger.info(f"El usuario {user_id} ha solicitado LINKS.")
-    chat_id = message.chat.id
-    bot.send_message(chat_id, get_links_message())
-
-
-@bot.message_handler(commands=['siu', 'campus'])
-def request_url_information(message):
-    user_id = message.from_user.id
-    chat_id = message.chat.id
-    logger.info(
-        f"El usuario {user_id} ha solicitado información de {message.text}.")
-    if message.text == "/siu":
-        url = 'https://autogestion.uno.edu.ar/uno/'
-        name = "siu guarani"
-    elif message.text == "/campus":
-        url = 'http://campusvirtual.uno.edu.ar/moodle/'
-        name = 'campus'
-
-    bot.send_message(
-        chat_id, f"<i>Solicitando información a {url} ...</i>")
-    bot.send_message(chat_id, url_message(url, name))
-
-
-@bot.message_handler(commands=['comunidades_it', 'comunidades'])
-def get_comunidades_it(message):
-    user_id = message.from_user.id
-    chat_id = message.chat.id
-    logger.info(f"El usuario {user_id} ha solicitado Comunidades IT.")
-    bot.send_message(chat_id, responses.comunidades_it())
-
-
-@bot.message_handler(commands=['calendar','calendario'])
-def get_academic_calendar(message):
-    arguments = message.text.split()
-    user_id = message.from_user.id
-    chat_id = message.chat.id
-    if len(arguments) > 1 and arguments[1] == 'feriados':
-        logger.info(f"El usuario {user_id} ha solicitado el Calendario de Feriados.")
-        bot.send_message(chat_id, responses.calendario_feriados_message())
-    else:
-        logger.info(f"El usuario {user_id} ha solicitado el Calendario Académico.")
-        bot.send_message(chat_id, responses.calendario_academico_message())
-
-
-@bot.message_handler(commands=['mails'])
-def get_emails(message):
-    arguments = message.text.split()
-    user_id = message.from_user.id
-    chat_id = message.chat.id
-    logger.info(f"El usuario {user_id} ha solicitado MAILS.")
-    if len(arguments) > 1:
-        bot.send_message(chat_id, responses.get_mails_by_term(arguments[1]))
-    else:
-        bot.send_message(chat_id, responses.get_mails_by_term())
+@bot.message_handler(commands=accepted_commands)
+def general_handler(message):
+    handlers[message.text.replace('/', '')](message)
 
 
 @server.route('/' + API_TOKEN, methods=['POST'])
